backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.0
python-dotenv==1.0.1
xarray==2024.06.0
netCDF4==1.7.1
numpy==2.0.1
pandas==2.2.2
copernicusmarine==2.1.2
aiofiles==24.1.0
